refactored how to determine whehter to process payee as an employee
